requirements.txt
numpy==2.1.1
omegaconf==2.3.0
PyYAML==6.0.1
PyYAML==6.0.2
torch==2.4.0
tqdm==4.66.5
